Use gint64 instead of long to be sure of the size
The size of type `long` can vary on different platforms. In addition,
the size of type `time_t` can vary as well. So be sure to cast `tv_sec`
to a known size.

On most platforms, `sizeof(time_t) == sizeof(gint64)`, so we should be
fine, but let's be safe.

+++ mono/utils/mono-time.c
@@ -104,7 +104,7 @@ inline void test_time_and_possibly_refresh_cache(struct timeval* now)
 
 	timersub(now, &last_time, &difference);
 
-	long diff_microseconds = (difference.tv_sec * 1e6) + difference.tv_usec;
+	gint64 diff_microseconds = ((gint64)difference.tv_sec * 1e6) + difference.tv_usec;
 
 	if (diff_microseconds > max_delta_time_jump_before_cache_invalidation_us|| diff_microseconds < 0)
 	{
